Create ChooseFontsPlugin.py
ChooseFontsPlugin adds the font menu item to your application that is using dearpygui as the GUI. Just insert the code  myFontsPlugin = EditFontsPlugin() in your dpg.menu_bar or dpg.viewport_menu_bar.

This plugin enables you to change the global font style of your dearpygui application. If you don't have a Fonts folder, the plugin will create one for you, as well as the files that will save the user's font preferences.